refact: Make separate video contols components
