+++ src/app/watch/memo.md
@@ -3,3 +3,4 @@
 - full screen 됐을 때 기본 내장 controls 적용되는 것 고치기
 - 음소거 버튼 만들기
 - 화면 눌렀을 때 재생/일시정지 되게 구현하기
+- material ui 적용한 css 아직
