+++ src/components/Watch/FullScreen/FullScreen.tsx
@@ -1,0 +1,50 @@
+import "./FullScreen.css";
+
+interface FullScreenProps {
+  videoRef: React.RefObject<HTMLVideoElement>;
+  isFullScreen: boolean;
+}
+
+function FullScreen({ videoRef, isFullScreen }: FullScreenProps) {
+  const handleFullScreenToggle = () => {
+    if (videoRef.current) {
+      if (!isFullScreen) {
+        // Enter full screen
+        if (videoRef.current.requestFullscreen) {
+          videoRef.current.requestFullscreen();
+        } else if (videoRef.current.mozRequestFullScreen) {
+          // Firefox
+          videoRef.current.mozRequestFullScreen();
+        } else if (videoRef.current.webkitRequestFullscreen) {
+          // Chrome, Safari and Opera
+          videoRef.current.webkitRequestFullscreen();
+        } else if (videoRef.current.msRequestFullscreen) {
+          // IE/Edge
+          videoRef.current.msRequestFullscreen();
+        }
+      } else {
+        // Exit full screen
+        if (document.exitFullscreen) {
+          document.exitFullscreen();
+        } else if (document.mozCancelFullScreen) {
+          // Firefox
+          document.mozCancelFullScreen();
+        } else if (document.webkitExitFullscreen) {
+          // Chrome, Safari and Opera
+          document.webkitExitFullscreen();
+        } else if (document.msExitFullscreen) {
+          // IE/Edge
+          document.msExitFullscreen();
+        }
+      }
+    }
+  };
+
+  return (
+    <button className="fullscreen-button" onClick={handleFullScreenToggle}>
+      {isFullScreen ? "Exit Full Screen" : "Full Screen"}
+    </button>
+  );
+}
+
+export default FullScreen;
