+++ src/components/Watch/PlayPause/PlayPause.css
@@ -1,0 +1,7 @@
+.play-pause-button {
+  background-color: rgba(0, 0, 0, 0.6);
+  color: white;
+  border: none;
+  padding: 10px 15px;
+  cursor: pointer;
+}
