+++ src/components/Watch/Player/Player.css
@@ -1,0 +1,16 @@
+/* components/Video.css */
+
+.video-player {
+  position: relative;
+}
+
+video {
+  width: 100%;
+}
+
+.controls {
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 10px;
+}
