+++ src/components/Watch/Skip/Skip.css
@@ -1,0 +1,11 @@
+.skip-container {
+  display: flex;
+}
+
+.skip-button {
+  background-color: rgba(0, 0, 0, 0.6);
+  color: white;
+  border: none;
+  padding: 10px 15px;
+  cursor: pointer;
+}
