+++ src/components/Watch/TimeBar/TimeBar.css
@@ -1,0 +1,15 @@
+.time-bar {
+  width: 100%;
+  height: 10px;
+  background-color: #ccc; /* Background color of the time bar container */
+  position: relative;
+}
+
+.time-bar-fill {
+  height: 100%;
+  background-color: #007bff; /* Color of the time bar fill */
+  width: 0; /* The width will be set dynamically based on time */
+  position: absolute;
+  top: 0;
+  left: 0;
+}
