--- src/components/Watch/Video.css
@@ -1,81 +1,0 @@
-/* components/Video.css */
-
-.video-player {
-  position: relative;
-}
-
-video {
-  width: 100%;
-}
-
-.controls {
-  display: flex;
-  align-items: center;
-  justify-content: space-between;
-  padding: 10px;
-}
-
-.play-pause-button {
-  background-color: rgba(0, 0, 0, 0.6);
-  color: white;
-  border: none;
-  padding: 10px 15px;
-  cursor: pointer;
-}
-
-.skip-button {
-  background-color: rgba(0, 0, 0, 0.6);
-  color: white;
-  border: none;
-  padding: 10px 15px;
-  cursor: pointer;
-}
-
-.progress-bar {
-  width: 100%;
-  height: 10px;
-  background-color: #ccc; /* Background color of the progress bar container */
-  position: relative;
-}
-
-.progress-bar-fill {
-  height: 100%;
-  background-color: #007bff; /* Color of the progress bar fill */
-  width: 0; /* The width will be set dynamically based on progress */
-  position: absolute;
-  top: 0;
-  left: 0;
-}
-
-.volume-control {
-  flex-grow: 1;
-  padding: 0 10px;
-}
-
-/* Style the volume slider */
-input[type="range"] {
-  width: 100%;
-  height: 10px;
-  -webkit-appearance: none;
-  appearance: none;
-  background: #ccc;
-  border-radius: 5px;
-  outline: none;
-  opacity: 0.7;
-  -webkit-transition: 0.2s;
-  transition: opacity 0.2s;
-}
-
-input[type="range"]::-webkit-slider-thumb {
-  -webkit-appearance: none;
-  appearance: none;
-  width: 15px;
-  height: 15px;
-  border-radius: 50%;
-  background: #007bff;
-  cursor: pointer;
-}
-
-input[type="range"]:hover {
-  opacity: 1;
-}
