+++ src/components/Watch/Volume/Volume.css
@@ -1,0 +1,32 @@
+.volume-control {
+  flex-grow: 1;
+  padding: 0 10px;
+}
+
+/* Style the volume slider */
+input[type="range"] {
+  width: 100%;
+  height: 10px;
+  -webkit-appearance: none;
+  appearance: none;
+  background: #ccc;
+  border-radius: 5px;
+  outline: none;
+  opacity: 0.7;
+  -webkit-transition: 0.2s;
+  transition: opacity 0.2s;
+}
+
+input[type="range"]::-webkit-slider-thumb {
+  -webkit-appearance: none;
+  appearance: none;
+  width: 15px;
+  height: 15px;
+  border-radius: 50%;
+  background: #007bff;
+  cursor: pointer;
+}
+
+input[type="range"]:hover {
+  opacity: 1;
+}
